Fix several msvc warnings

+++ samples/binary_search/binary_search.cpp
@@ -265,7 +265,7 @@ ForwardIterator simd_lower_bound2( ForwardIterator beg, ForwardIterator end,
     // Create SIMD index key
     ls::simd_type< int32_t, TAG_T > indexes =
         ls::mullo( ls::simd_type< int32_t, TAG_T >( 8, 7, 6, 5, 4, 3, 2, 1 ),
-                   ls::simd_type< int32_t, TAG_T >( step ) );
+                   ls::simd_type< int32_t, TAG_T >( static_cast<int32_t>( step ) ) );
 
     // Create SIMD search key
     simd_type cmp = i32Gather< TAG_T >( start, indexes );
